[Refactor] Move StarOS integration into ComputeEnv
