+++ be/module_boundary_manifest.json
@@ -455,11 +455,12 @@
     {
       "id": "computeenv",
       "doc_label": "ComputeEnv",
-      "summary": "Shared compute-side BE environment boundary for process-scoped compute resources, load-path management, query-scoped scan coordination helpers, query-cache primitives, WorkGroup scheduling/executor resources, and pipeline controls below full Exec/Storage and above RuntimeEnv.",
+      "summary": "Shared compute-side BE environment boundary for process-scoped compute resources, StarOS/Starlet worker runtime, load-path management, query-scoped scan coordination helpers, query-cache primitives, WorkGroup scheduling/executor resources, StarCache coordination, and pipeline controls below full Exec/Storage and above RuntimeEnv.",
       "owned_targets": ["ComputeEnv"],
       "owned_roots": ["be/src/compute_env"],
       "allowed_include_prefixes": [
         "compute_env/",
+        "cache/",
         "exec/runtime_filter/",
         "exec/pipeline/pipeline_fwd.h",
         "exec/pipeline/operator.h",
@@ -490,10 +491,10 @@
       ],
       "forbidden_include_prefixes": ["exec/", "storage/", "service/", "connector/", "agent/"],
       "forbidden_includes": ["runtime/exec_env.h"],
-      "allowed_target_deps": ["StoragePrimitive", "ExecPrimitive", "ExprCore", "Serde", "RuntimeEnv", "RuntimeCore", "Platform", "FSCore", "IO", "ColumnSortCore", "ChunkCore", "ColumnCore", "Types", "Common", "Base", "Gutil", "StarRocksGen"],
-      "allowed_test_targets": ["compute_env_test", "compute_env_sorting_test"],
+      "allowed_target_deps": ["Cache", "StoragePrimitive", "ExecPrimitive", "ExprCore", "Serde", "RuntimeEnv", "RuntimeCore", "Platform", "FSCore", "IO", "ColumnSortCore", "ChunkCore", "ColumnCore", "Types", "Common", "Base", "Gutil", "StarRocksGen"],
+      "allowed_test_targets": ["compute_env_test", "compute_env_sorting_test", "compute_env_staros_test"],
       "allowed_test_link_deps": ["ComputeEnv", "ExprCore", "RuntimeCore", "FSCore", "IO", "ChunkCore", "ColumnCore", "Types", "Common", "Base", "Gutil", "StarRocksGen"],
-      "remediation": "Keep ComputeEnv limited to process-scoped compute resources, load-path management without ExecEnv or concrete storage engine coupling, query-scoped scan coordination helpers without concrete scan/storage policy, query-cache primitives without concrete storage policy, WorkGroup scheduling/executor resources, shared compute-side service contracts, stable execution primitives, reusable compute-side sorting algorithms, and spill infrastructure; move concrete Exec, Storage, Service, Connector, and Agent integration upward."
+      "remediation": "Keep ComputeEnv limited to process-scoped compute resources, StarOS/Starlet worker runtime, load-path management without ExecEnv or concrete storage engine coupling, query-scoped scan coordination helpers without concrete scan/storage policy, query-cache primitives without concrete storage policy, WorkGroup scheduling/executor resources, shared compute-side service contracts, stable execution primitives, reusable compute-side sorting algorithms, StarCache coordination, and spill infrastructure; move concrete Exec, Storage, Service, Connector, and Agent integration upward."
     },
     {
       "id": "agentserver",
@@ -1017,18 +1018,6 @@
       "remediation": "Keep ExecJoinCore limited to reusable join hash table algorithms; leave join nodes, pipeline operators, storage/service integration, and util diagnostics in higher modules."
     },
     {
-      "id": "starosintegration",
-      "doc_label": "StarOSIntegration",
-      "summary": "StarRocks-owned Starlet integration for worker runtime, shard metadata, status conversion, StarCache updates, and worker metrics outside service bootstrap.",
-      "owned_targets": ["StarOSIntegration"],
-      "owned_roots": ["be/src/staros_integration"],
-      "allowed_include_prefixes": ["staros_integration/", "common/", "base/", "gutil/", "gen_cpp/"],
-      "allowed_target_deps": ["Common", "Base", "Gutil", "StarRocksGen"],
-      "allowed_test_targets": ["staros_integration_test"],
-      "allowed_test_link_deps": ["StarOSIntegration", "Common", "Base", "Gutil", "StarRocksGen"],
-      "remediation": "Keep the Starlet adapter independent of service/bootstrap, storage, runtime, and filesystem modules; expose narrow accessors instead of pulling higher layers into lower layers."
-    },
-    {
       "id": "service",
       "doc_label": "Service",
       "summary": "Shared service-layer target above runtime, cache, compute, and AgentServer without owning ServiceBE bootstrap code.",
@@ -1053,7 +1042,6 @@
         "platform/",
         "runtime/",
         "serde/",
-        "staros_integration/",
         "storage/",
         "types/",
         "util/"
@@ -1081,7 +1069,6 @@
         "Base",
         "Gutil",
         "StarRocksGen",
-        "StarOSIntegration",
         "Connector",
         "ConnectorPrimitive",
         "ConnectorBootstrap",
