+++ be/src/compute_env/staros/staros_worker.h
