+++ be/src/http/action/stop_be_action.cpp
@@ -25,7 +25,7 @@
 #include "http/http_status.h"
 
 #ifdef USE_STAROS
-#include "staros_integration/staros_worker_runtime.h"
+#include "compute_env/staros/staros_worker_runtime.h"
 #endif
 
 namespace starrocks {
