+++ be/src/service/backend_metrics_initializer.cpp
@@ -35,6 +35,7 @@
 #include "common/status.h"
 #include "common/system/backend_options.h"
 #include "common/system/disk_info.h"
+#include "compute_env/staros/staros_worker_metrics.h"
 #include "exec/pipeline/primitives/pipeline_metrics.h"
 #include "fs/fs.h"
 #include "fs/fs_util.h"
@@ -45,7 +46,6 @@
 #include "runtime/runtime_metrics.h"
 #include "runtime/stream_load/stream_load_metrics.h"
 #include "service/service_metrics.h"
-#include "staros_integration/staros_worker_metrics.h"
 #include "storage/storage_metrics.h"
 #ifndef __APPLE__
 #include "util/jvm_metrics.h"
