--- be/test/staros_integration/CMakeLists.txt
@@ -1,52 +1,0 @@
-# Copyright 2021-present StarRocks, Inc. All rights reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     https://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-set(BE_TEST_MACOS_LINK_STUBS)
-if (APPLE)
-    list(APPEND BE_TEST_MACOS_LINK_STUBS ../base/macos_link_stubs.cpp)
-endif()
-
-set(STAROS_INTEGRATION_TEST_FILES
-    ${BE_TEST_MACOS_LINK_STUBS}
-    staros_worker_metrics_test.cpp
-)
-
-if (NOT APPLE)
-    list(APPEND STAROS_INTEGRATION_TEST_FILES ../base/cxa_throw_wrap.cpp)
-endif()
-
-if ("${USE_STAROS}" STREQUAL "ON")
-    list(APPEND STAROS_INTEGRATION_TEST_FILES staros_worker_test.cpp)
-endif()
-
-set(STAROS_INTEGRATION_TEST_LINK_LIBS
-    StarOSIntegration
-    Common
-    Base
-    Gutil
-    StarRocksGen
-    ${STARROCKS_DEPENDENCIES}
-    ${STARROCKS_STATIC_LINK_LIBS}
-    ${WL_LINK_DYNAMIC} ${STARROCKS_SYSTEM_LINK_LIBS} ${WRAP_LINKER_FLAGS}
-    ${WL_START_GROUP}
-    gmock
-    gtest
-    ${WL_END_GROUP}
-)
-
-add_library(starrocks_staros_integration_test_objs OBJECT ${STAROS_INTEGRATION_TEST_FILES})
-target_link_libraries(starrocks_staros_integration_test_objs PRIVATE StarRocksPCH)
-set_target_properties(starrocks_staros_integration_test_objs PROPERTIES COMPILE_FLAGS "-fno-access-control")
-add_executable(staros_integration_test $<TARGET_OBJECTS:starrocks_staros_integration_test_objs>)
-target_link_libraries(staros_integration_test ${STAROS_INTEGRATION_TEST_LINK_LIBS} gtest_main)
